requirements.txt
langchain-aws>=0.1.0
langchain-core>=0.2.0
langchain>=0.2.0
boto3>=1.34.0
python-dotenv>=1.0.0
pytest>=8.0.0
pytest-cov>=4.1.0
black>=24.0.0
flake8>=7.0.0
# For potential future use
# langchain-community>=0.0.10
# langsmith>=0.0.70
# langchain-openai>=0.0.5